Fix for page flushing
The store now ensures a flush to disk by stopping and restarting the
background page writer. This appears to be required as a simple
Stream.Flush() does not always flush to disk. This fixes a problem where
querying a store in Polaris after an import has run on the store can
sometimes lead to "resource not found" errors.