Workaround for issue #15
The hang on parsing only occurs when using the
AsynchronousBufferDuringParsing token queue mode in dotNetRDF. This fix
changes that to a SynchronousBufferDuringParsing which requires more
memory but does not exhibit the hang.
Also removed the code that created the decompressing GZipStream wrapper
as this is done inside the dotNetRDF library.