Fix concurrent stream access issues on WinRT
Provides a workaround for the oplock issues under WinRT by providing a
Stream implementation that automatically reopens itself when an oplock
error occurs. This fix also requires disabling the background writer
thread under WinRT and so it also extends the API for
BinaryFilePageStore to allow the use of background page write to be
specified in the constructor.