Fixing up CORS in errors
See issue #210
Fixed a problem with an invalid store name in the SPARQL query route
resulting in a 500 rather than a 404 error.
Added code to the Nancy error pipeline to detect a 500 error on a
request that accepts JSON and return a serialization of the exception
along and set CORS headers.