Fixes for T4 NuGet package
Fixed T4 Template to ensure that project and solution files are saved
before invoking the Roslyn process.
BrightstarDB assembly is now located using project metadata and the
generator version number is generated from the assembly version of the
generator assembly instead of BrightstarDB. These changes mean that the
BrightstarDB.dll does not need to be included in the tools directory of
the package.